Rewrite previously excluded raw_ptr fields
This rewrites pointers to byte buffers that were accidentally skipped by
the rewriter. In addition to skipping string literals, the rewriter was
wrongly skipping byte buffers.

DanglingUntriaged-notes: Annotating pre-existing dangling pointers.
Bug: 331840473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5497157
Owners-Override: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Ali Hijazi <ahijazi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1306660}

+++ media/filters/memory_data_source.cc
@@ -39,7 +39,7 @@ void MemoryDataSource::Read(int64_t position,
 
   if (clamped_size > 0) {
     DCHECK(data);
-    memcpy(data, data_ + position, clamped_size);
+    memcpy(data, data_ + base::checked_cast<size_t>(position), clamped_size);
   }
 
   std::move(read_cb).Run(clamped_size);
